refactor(codegen): extract runtime-ABI call lowering into runtime_abi.rs
Relocate lower_call_runtime_abi — the ~2.5k-line single dispatch that lowers
every Instr::CallRuntimeAbi across the whole RuntimeCallFamily surface — out of
the llvm.rs god-module, together with the two bytes-ABI consumer-convention
predicates (is_bytes_by_pointer_consumer, is_bytes_struct_expansion_consumer)
that the catalog FFI declaration edge and the call edge both consult.

Widens the shared call_runtime/_void/_basic/_int/_ptr FnCtx helpers plus the
per-family emit helpers the dispatch calls to pub(crate). Pure relocation
following the coro.rs carve; llvm.rs drops ~2.5k lines and every emitted .ll is
byte-identical (ll-corpus verify clean, native and wasm32).
