+++ hew-codegen-rs/src/lib.rs
@@ -32,6 +32,7 @@
 pub(crate) mod arith;
 pub mod coro;
 pub mod llvm;
+pub(crate) mod runtime_abi;
 
 pub use llvm::{
     emit_module, emit_module_objects, validate_codegen_front, verify_pipeline, CodegenError,
